+++ test/oauth2/authorize/http/response/codetoken.test.js
@@ -47,4 +47,111 @@ describe('oauth2/authorize/http/response/codetoken', function() {
       .catch(done);
   }); // should create response type without response modes
   
+  describe('issueToken', function() {
+    var container = new Object();
+    container.components = sinon.stub()
+    container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseMode').returns([]);
+    container.components.withArgs('http://i.authnomicon.org/oauth2/authorization/http/ResponseParameters').returns([]);
+    var ats = new Object();
+    
+    var codeTokenSpy = sinon.stub();
+    var factory = $require('../../../../../com/oauth2/authorize/http/response/codetoken', {
+      'oauth2orize-openid': {
+        grant: { codeToken: codeTokenSpy }
+      }
+    });
+    
+    var issueToken;
+    
+    beforeEach(function(done) {
+      ats.issue = sinon.stub().yieldsAsync(null, '2YotnFZFEjr1zCsicMWpAA');
+      
+      factory(ats, null, logger, container)
+        .then(function(type) {
+          issue = codeTokenSpy.getCall(0).args[1];
+          done();
+        })
+        .catch(done);
+    });
+    
+    it('should issue access token', function(done) {
+      var client = {
+        id: 's6BhdRkqt3',
+        name: 'My Example Client'
+      };
+      var user = {
+        id: '248289761001',
+        displayName: 'Jane Doe'
+      };
+      var ares = {
+        allow: true
+      }
+      var areq = {
+        type: 'code',
+        clientID: 's6BhdRkqt3',
+        redirectURI: 'https://client.example.com/cb',
+        state: 'xyz'
+      }
+      
+      issue(client, user, ares, areq, {}, function(err, token) {
+        if (err) { return done(err); }
+        
+        expect(ats.issue.callCount).to.equal(1);
+        expect(ats.issue.getCall(0).args[0]).to.deep.equal({
+          client: {
+            id: 's6BhdRkqt3',
+            name: 'My Example Client'
+          },
+          user: {
+            id: '248289761001',
+            displayName: 'Jane Doe'
+          }
+        });
+        expect(token).to.equal('2YotnFZFEjr1zCsicMWpAA');
+        done();
+      });
+    }); // should issue access token
+    
+    it('should issue access token with scope', function(done) {
+      var client = {
+        id: 's6BhdRkqt3',
+        name: 'My Example Client'
+      };
+      var user = {
+        id: '248289761001',
+        displayName: 'Jane Doe'
+      };
+      var ares = {
+        allow: true,
+        scope: [ 'profile', 'email' ]
+      }
+      var areq = {
+        type: 'code',
+        clientID: 's6BhdRkqt3',
+        redirectURI: 'https://client.example.com/cb',
+        state: 'xyz'
+      }
+      
+      issue(client, user, ares, areq, {}, function(err, token) {
+        if (err) { return done(err); }
+        
+        expect(ats.issue.callCount).to.equal(1);
+        expect(ats.issue.getCall(0).args[0]).to.deep.equal({
+          client: {
+            id: 's6BhdRkqt3',
+            name: 'My Example Client'
+          },
+          user: {
+            id: '248289761001',
+            displayName: 'Jane Doe'
+          },
+          scope: [ 'profile', 'email' ]
+        });
+        expect(token).to.equal('2YotnFZFEjr1zCsicMWpAA');
+        done();
+      });
+    }); // should issue access token with scope
+    
+  }); // issueToken
+  
 });