+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- `AccessTokenService#issue()` is called with `msg` argument containing `scope`
+property, rather than `grant.scope`, when called from `code token` response
+type.
 
 ## [0.0.3] - 2021-12-02
 ### Added